Implemented Hash
The hash function now takes two parameters. The main function initializes the hashSize variable with a value (in this case, 10). The code creates the hash table by dynamically allocating memory for the struct HashType. It then reads the records from "input.txt" using the parseData function and stores them in an array of struct RecordType. The insertRecord function is used to insert each record into the hash table using separate chaining. after inserting all records into the hash table, the displayRecordsInHash function is called to print the contents. Lastly, the code frees the memory.